add moe module
